Create wordBreak.cpp
This question is actually a question of dp but can be done using recursion too with more time complexity. If you try to run it on leetcode it'll show time limit exceeded. The more efficient solution is can be done using dp.
Given a string s and a dictionary of strings wordDict, return true if s can be segmented into a space-separated sequence of one or more dictionary words.
Note that the same word in the dictionary may be reused multiple times in the segmentation.